pyproject.toml
[project]
name = "deepfix"
version = "0.1.0"
description = "DeepFix is a copilot designed to provide systematic guidance for machine learning model training, and debugging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anywidget>=0.9.18",
    "datasets>=4.0.0",
    "deepchecks[vision]>=0.19.1",
    "dspy>=3.0.3",
    "fastapi>=0.117.1",
    "fire>=0.7.1",
    "httpx>=0.28.1",
    "huggingface-hub[hf-xet]>=0.34.4",
    "kaleido>=1.1.0",
    "lightning>=2.5.4",
    "llama-index-retrievers-bm25>=0.6.5",
    "mlflow>=3.3.2",
    "notebook>=7.4.5",
    "omegaconf>=2.3.0",
    "open-clip-torch",
    "pandas>=2.2.3",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "seaborn>=0.13.2",
    "sqlmodel>=0.0.24",
    "structlog>=25.4.0",
    "timm>=1.0.19",
    "torch>=2.0.0",
    "torchmetrics",
    "torchvision",
    "typer>=0.16.1",
]

[project.scripts]
deepfix = "deepfix.cli.main:main"
